workaround for race between waitpid(-1) and waitpid(pid)
Sometimes waitpid(-1) in cb_check_data reaps a child that has just terminated,
before waitpid(pid) in wait_process can reap it.  Linux waitpid(pid,,WNOHANG)
may not return an error if the child has already terminated, which results in
wait_process looping waiting for the child which has already terminated.  The
symptom is a spurious "timeout waiting for filter command to exit".